Move CreateEntitySetModal
