+++ src/components/data/EntitiesContent.js
@@ -10,7 +10,7 @@ import { tools } from 'src/analysis/utils/tool-utils';
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils';
 import { ButtonSecondary } from 'src/components/common';
 import { DataTableColumnProvenance } from 'src/components/data/data-table-provenance';
-import { CreateEntitySetModal, ModalToolButton } from 'src/components/data/data-utils';
+import { ModalToolButton } from 'src/components/data/data-utils';
 import DataTable from 'src/components/data/DataTable';
 import ExportDataModal from 'src/components/data/ExportDataModal';
 import { icon, spinner } from 'src/components/icons';
@@ -24,6 +24,7 @@ import TitleBar from 'src/components/TitleBar';
 import WorkflowSelector from 'src/components/WorkflowSelector';
 import { AddColumnModal } from 'src/data/data-table/entity-service/AddColumnModal';
 import { AddEntityModal } from 'src/data/data-table/entity-service/AddEntityModal';
+import { CreateEntitySetModal } from 'src/data/data-table/entity-service/CreateEntitySetModal';
 import { entityAttributeText } from 'src/data/data-table/entity-service/entityAttributeText';
 import { EntityDeleter } from 'src/data/data-table/entity-service/EntityDeleter';
 import { MultipleEntityEditor } from 'src/data/data-table/entity-service/MultipleEntityEditor';
