+++ src/components/data/data-utils.js
@@ -1,18 +1,9 @@
 import _ from 'lodash/fp';
 import { Fragment, useRef, useState } from 'react';
-import { b, div, h, img, label } from 'react-hyperscript-helpers';
-import {
-  absoluteSpinnerOverlay,
-  ButtonPrimary,
-  Clickable,
-  DeleteConfirmationModal,
-  IdContainer,
-  Link,
-  Select,
-  spinnerOverlay,
-} from 'src/components/common';
+import { b, div, h, img } from 'react-hyperscript-helpers';
+import { absoluteSpinnerOverlay, ButtonPrimary, Clickable, DeleteConfirmationModal, Link, Select, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
-import { ConfirmedSearchInput, TextInput } from 'src/components/input';
+import { ConfirmedSearchInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import Modal from 'src/components/Modal';
 import PopupTrigger, { MenuDivider } from 'src/components/PopupTrigger';
@@ -26,7 +17,6 @@ import { reportError } from 'src/libs/error';
 import { notify } from 'src/libs/notifications';
 import { requesterPaysProjectStore } from 'src/libs/state';
 import * as Style from 'src/libs/style';
-import * as Utils from 'src/libs/utils';
 
 export const warningBoxStyle = {
   backgroundColor: colors.warning(0.15),
@@ -151,107 +141,6 @@ export const notifyDataImportProgress = (jobId, message) => {
     id: jobId,
     message,
   });
-};
-
-export const CreateEntitySetModal = ({ entityType, entityNames, workspaceId: { namespace, name: workspaceName }, onDismiss, onSuccess }) => {
-  const [name, setName] = useState('');
-  const [nameInputTouched, setNameInputTouched] = useState(false);
-  const nameError =
-    nameInputTouched &&
-    Utils.cond(
-      [!name, () => 'A name for the set is required.'],
-      [!/^[A-Za-z0-9_-]+$/.test(name), () => 'Set name may only contain alphanumeric characters, underscores, dashes, and periods.']
-    );
-
-  const [isBusy, setIsBusy] = useState();
-
-  const createSet = async () => {
-    setIsBusy(true);
-    try {
-      await Ajax()
-        .Workspaces.workspace(namespace, workspaceName)
-        .createEntity({
-          name,
-          entityType: `${entityType}_set`,
-          attributes: {
-            [`${entityType}s`]: {
-              itemsType: 'EntityReference',
-              items: _.map((entityName) => ({ entityType, entityName }), entityNames),
-            },
-          },
-        });
-      onSuccess();
-    } catch (e) {
-      onDismiss();
-      reportError('Unable to create set.', e);
-    }
-  };
-
-  return h(
-    Modal,
-    {
-      title: `Create a ${entityType} set`,
-      onDismiss,
-      okButton: h(
-        ButtonPrimary,
-        {
-          disabled: !name || nameError,
-          tooltip: nameError,
-          onClick: createSet,
-        },
-        ['Save']
-      ),
-    },
-    [
-      div({ style: { display: 'flex', flexDirection: 'column', marginBottom: '1rem' } }, [
-        h(IdContainer, [
-          (id) =>
-            h(Fragment, [
-              label({ htmlFor: id, style: { fontWeight: 'bold', marginBottom: '0.5rem' } }, 'Set name (required)'),
-              div({ style: { position: 'relative', display: 'flex', alignItems: 'center' } }, [
-                h(TextInput, {
-                  id,
-                  value: name,
-                  placeholder: 'Enter a name for the set',
-                  style: nameError
-                    ? {
-                        paddingRight: '2.25rem',
-                        border: `1px solid ${colors.danger()}`,
-                      }
-                    : undefined,
-                  onChange: (value) => {
-                    setName(value);
-                    setNameInputTouched(true);
-                  },
-                }),
-                nameError &&
-                  icon('error-standard', {
-                    size: 24,
-                    style: {
-                      position: 'absolute',
-                      right: '0.5rem',
-                      color: colors.danger(),
-                    },
-                  }),
-              ]),
-              nameError &&
-                div(
-                  {
-                    'aria-live': 'assertive',
-                    'aria-relevant': 'all',
-                    style: {
-                      marginTop: '0.5rem',
-                      color: colors.danger(),
-                    },
-                  },
-                  nameError
-                ),
-            ]),
-        ]),
-      ]),
-      isBusy && spinnerOverlay,
-    ]
-  );
 };
 
 export const ModalToolButton = ({ icon, text, disabled, ...props }) => {
